update Cping class to use chrono lib

+++ Cping.h
@@ -10,5 +10,5 @@ class Cping : Cgpio
 	
 	Cping(int gpio);
 	~Cping();	
-	int getDistance();
+	float getDistance();
 };
